Support configuring elasticsearch-curator version
Adds an elasticsearch_curator_version variable to support defining which
version to install and defaults it to version 3.5.1 since the role
currently doesn't support the new yml configuration file required in the
4.* releases.

Also adds tasks to register the python version on the host and use that
to decide if the argparse pip library needs to be installed as a
dependency. It's no longer a requirement to install that dependency for
more recent Python versions because:

"As of Python >= 2.7 and >= 3.2, the argparse module is maintained
within the Python standard library. For users who still need to support
Python < 2.7 or < 3.2" - https://pypi.python.org/pypi/argparse